% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expanded_data.R
\docType{data}
\name{expanded_data}
\alias{expanded_data}
\title{RT-qPCR ΔCT Data for 406 Endometrial Tissue Biopsies.}
\format{
A data frame with 257 rows and 9 variables.
\describe{
\item{ID}{Unique biopsy identification}
\item{BATCH}{The batch number to which the sample contributed}
\item{LH}{The number of days since ovulation, as provided by the donating patient based on the results of urinary ovulation test}
\item{CXCL14}{-ΔCT values for CXCL14}
\item{DPP4}{-ΔCT values for DPP4}
\item{GPX3}{-ΔCT values for GPX3}
\item{IGFBP1}{-ΔCT values for IGFBP1}
\item{IL2RB}{-ΔCT values for IL2RB}
\item{SLC15A2}{-ΔCT values for SLC15A2}
}
}
\usage{
expanded_data
}
\description{
A dataset comprised of the results of RT-qPCR analysis for 406 endometrial tissue
biopsies, with associated metadata. L19 has been utilised as a housekeeping gene.
Batch correction has been performed and ΔCT values inverted to -ΔCT to reflect
expression.
}
\keyword{datasets}
