% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endotime_train.R
\name{endotime_train}
\alias{endotime_train}
\title{Title}
\usage{
endotime_train(
  expression_data,
  genes,
  batch_correct = TRUE,
  check_asynchrony = TRUE,
  gene_window = 80,
  aggregate_window = 80,
  euc = 2,
  enforce_monotony = FALSE,
  decreasing_window = FALSE,
  new_extrap = FALSE,
  convert_realtime = TRUE
)
}
\arguments{
\item{expression_data}{...}

\item{genes}{...}

\item{batch_correct}{...}

\item{check_asynchrony}{...}

\item{gene_window}{...}

\item{aggregate_window}{...}

\item{euc}{...}

\item{enforce_monotony}{...}

\item{decreasing_window}{...}

\item{new_extrap}{...}

\item{convert_realtime}{...}
}
\value{
...
}
\description{
Title
}
